IBM-Swift/Kitura#505 extracted FileResourceServer class
put sendIfFound() and supporting functions there
